requirements.txt
# Chic Homz — Agentic Sales Agent Dependencies

# API server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# LangGraph agent framework
langgraph>=0.2.0
langgraph-cli[inmem]

# LangChain ecosystem
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-huggingface
langchain-pinecone
langchain-community

# Embeddings & vector stores
sentence-transformers
faiss-cpu
pinecone-client>=3.0.0

# Utilities
python-dotenv>=1.0.0
pydantic

# Testing
pytest
